split entrypoint for openmotics service
